Replace `notes` with `sb`
This is in light of my transition of having a generic set of notes to
building a second brain - this helps me to organise my notes and reduces
the barrier of entry when it comes to writing.

The `sb` script for all intents and purposes is the same as `notes` with
a few tweaks where I've made changes to the structure of notes e.g.
markdown links to wiki links to make moving files around the second
brain easier - allowing me to simply the logic used to extract the
filename from a link.